Code smells, move code to helpers
