+++ src/services/event.service.ts
@@ -1,7 +1,6 @@
 import {SupabaseClient} from "@supabase/supabase-js";
 import {Database} from "@/types/database";
 import {Attending} from "@/types/database-types";
-import {Event} from "@/types/database-types"
 import {IEvent} from "@/interfaces/event";
 
 export const getEvent = (supabase: SupabaseClient<Database>, id: string) => {
